极验3滑块/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==20.3.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.4
cssselect==1.1.0
ddddocr==1.4.3
environs==9.5.0
fake-headers==1.0.2
Flask==1.1.4
flatbuffers==2.0
frozenlist==1.3.0
gevent==21.12.0
greenlet==1.1.2
html5lib==1.1
idna==3.3
itsdangerous==0.24
Jinja2==2.11.3
loguru==0.6.0
lxml==4.9.0
MarkupSafe==2.0.1
marshmallow==3.16.0
maxminddb==2.2.0
maxminddb-geolite2==2018.703
multidict==6.0.2
numpy==1.22.4
onnxruntime==1.11.1
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.1
protobuf==4.21.1
pycparser==2.21
PyExecJS==1.5.1
pyparsing==3.0.9
pyquery==1.4.3
python-dotenv==0.20.0
redis==3.5.3
requests==2.27.1
retrying==1.3.3
six==1.16.0
soupsieve==2.3.2.post1
supervisor==4.2.4
tornado==6.1
urllib3==1.26.9
webencodings==0.5.1
Werkzeug==1.0.1
win32-setctime==1.1.0
yarl==1.7.2
zope.event==4.5.0
zope.interface==5.4.0
